make spatialite db readable also if no style is defined - in readonly mode

+++ geopaparazzispatialitelibrary/src/eu/geopaparazzi/spatialite/database/spatial/core/databasehandlers/SpatialiteDatabaseHandler.java
@@ -22,6 +22,7 @@ import android.graphics.Paint;
 import android.graphics.Paint.Cap;
 import android.graphics.Paint.Join;
 
+import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -233,7 +234,7 @@ public class SpatialiteDatabaseHandler extends AbstractSpatialDatabaseHandler {
         if (propertiesTableColumnCount == 0) {
             createPropertiesTable(dbJava);
             for (SpatialVectorTable spatialTable : vectorTableList) {
-                createDefaultPropertiesForTable(dbJava, spatialTable.getUniqueNameBasedOnDbFilePath(),
+                createDefaultPropertiesForTable(dbJava, spatialTable.getUniqueNameBasedOnTableName(),
                         spatialTable.getLabelField());
             }
         }
@@ -776,12 +777,16 @@ public class SpatialiteDatabaseHandler extends AbstractSpatialDatabaseHandler {
                 // assign the styles
                 for (SpatialVectorTable spatialTable : vectorTableList) {
                     Style style4Table = null;
+
+                    File parentFile = spatialTable.getDatabaseFile().getParentFile();
                     try {
-                        style4Table = getStyle4Table(dbJava, spatialTable.getUniqueNameBasedOnDbFilePath(),
+                        style4Table = getStyle4Table(dbJava, spatialTable.getUniqueNameBasedOnTableName(),
                                 spatialTable.getLabelField());
                     } catch (java.lang.Exception e) {
-                        deleteStyleTable(dbJava);
-                        checkPropertiesTable();
+                        if (parentFile.canWrite()) {
+                            deleteStyleTable(dbJava);
+                            checkPropertiesTable();
+                        }
                     }
                     if (style4Table == null) {
                         spatialTable.makeDefaultStyle();
@@ -814,7 +819,7 @@ public class SpatialiteDatabaseHandler extends AbstractSpatialDatabaseHandler {
         deleteStyleTable(dbJava);
         createPropertiesTable(dbJava);
         for (SpatialVectorTable spatialTable : vectorTableList) {
-            createDefaultPropertiesForTable(dbJava, spatialTable.getUniqueNameBasedOnDbFilePath(),
+            createDefaultPropertiesForTable(dbJava, spatialTable.getUniqueNameBasedOnTableName(),
                     spatialTable.getLabelField());
         }
     }
